알림 시스템 버그 수정 + 설정 탭 분리 + 토스트 애니메이션
1. 토스트 알림 미표시 근본 원인 수정:
   - updateCell/bulkUpdateCells에 updatedBy(senderId) 전달 누락 → 수정
   - updateSceneField에 senderId 파라미터 추가 (IPC 체인 전체)
   - senderId가 undefined여서 알림 조건 항상 false 되던 문제 해결

2. 댓글 알림 broadcast 추가:
   - comment-added broadcast 이벤트 신설 (postgres_changes 의존 제거)
   - App.tsx에서 comment-added 수신 → dispatchNotification 호출

3. 설정 탭에 알림 섹션 분리:
   - NotificationSection 컴포넌트 신규 생성
   - SettingsSidebar에 '알림' 탭 추가 (프로필 바로 아래)
   - ProfileSection에서 알림 관련 코드 제거 (중복 제거)
   - noti-settings-changed 이벤트로 App.tsx notiSettingsRef 실시간 동기화

4. 토스트 애니메이션 개선:
   - 스르륵 슬라이드 인 (cubic-bezier 이징)
   - 상단/하단 위치별 방향 분리
   - 호버 시 스택 부드럽게 펼침 (transition 0.35s)
   - 닫힐 때 페이드 아웃 애니메이션
